Remove Mixpanel integration
This project has been started by cloning one of the docs.resin.io pages
using HTTrack, which included all the code necessary to send events to
our production Mixpanel application.

I'm removing this code, as probably we don't need to track events on a
demo project, and even if so, we should not be using our production
Mixpanel token for it.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>